prevent tschenggins-status.pl processes accumulating on webserver
Save pid so that previous instances can terminate in case they're still running and haven't noticed yet that the Lämpli is gone.
Otherwise Apache might be waiting a looooong time to kill the CGI process (TCP timeout) as it doesn't know the Lämpli has a new connection (which happens on a reset or when the wifi reconnects or so).